+++ src/main/kotlin/com/mkotb/ktesla/request/commands/DoorRequests.kt
@@ -8,7 +8,7 @@ import io.ktor.http.HttpMethod
  * Unlocks the door to the car and extends handles if applicable
  */
 class UnlockDoorRequest (
-        vehicleId: Int
+        vehicleId: Long
 ) : TeslaRequest<SuccessResponse>() {
     override val endpoint = "api/1/vehicles/$vehicleId/command/door_unlock"
     override val method = HttpMethod.Post
@@ -18,7 +18,7 @@ class UnlockDoorRequest (
  * Locks the door to the car and detracts handles if applicable
  */
 class LockDoorRequest (
-        vehicleId: Int
+        vehicleId: Long
 ) : TeslaRequest<SuccessResponse>() {
     override val endpoint = "api/1/vehicles/$vehicleId/command/door_lock"
     override val method = HttpMethod.Post
