+++ src/main/kotlin/com/mkotb/ktesla/request/data/ClimateStateRequest.kt
@@ -6,7 +6,7 @@ import com.mkotb.ktesla.request.TeslaResponse
 import io.ktor.http.HttpMethod
 
 class ClimateStateRequest (
-        vehicleId: Int
+        vehicleId: Long
 ) : TeslaRequest<ClimateStateResponse>() {
     override val endpoint = "api/1/vehicles/$vehicleId/data_request/climate_state"
     override val method = HttpMethod.Get
